Update README.md
Line 17: "customizable conditions which requests to log" -- the language was a bit unclear. Is my edit correct? Should it be "conditions that make requests to log" or "conditions that make logging requests"?
Line 60:  "All integrations require an instance of Logbook which holds all configuration and wires all necessary parts together" does this mean that you need "an instance of Logbook that holds all configuration...," or does this mean that "Logbook [always] holds all configuration...?