Update README.md
Line 139: "usually" ... wondering if "otherwise" would be closer? normalerweise vs andernfalls vs sonst
Line 143: "*Formatting* defines how requests and responses will be transformed to strings basically. " should it be "strings, basically" or "strings [in a basic way]?"
Line 214: I.e., a JSON response body won’t be escaped and [or?] represented as a string -- wondering if it's "and" or "or". Also wondering if you might even mean this: "a JSON response body will **not** be escaped, BUT WILL [INSTEAD] represented as a string:"
LINE 237: "logged with an *slf4j* logger that uses the `org.zalando.logbook.Logbook` category" -- even for non-Zalandos?

copy replacement/edit done; now going to check section headers for consistency and add TOC