Fix: The underlying HTTP client completed without emitting a response (#1476)
* Fix: The underlying HTTP client completed without emitting a response

LogbookExchangeFilterFunction was not handing case of empty response with `Transfer-Encoding: chunked`. This should not happen, but it happens. This resulted in WebClient throwing an error as ExchangeFunction returned an empty Mono.

The issue was not reproducible without Wiremock as Spring would always return `Content-Length: 0` with empty response. To add tests for the issue, I rewrote LogbookExchangeFilterFunctionTest to use WireMock.

Note: I added wiremock version 3.0.0-beta-7, which supports jakarta servlet api. Versions lower than 3 were resulting in application startup errors.

* Cover both: content-length and chunked flows

* return test scope to submodules

* Check for Content-Length header to be present