Fix errors and warnings during build
maven-compiler-plugin 3.11.0 shows warnings by default. We have a
compilation warning as we use JDK 17 to compile, but
maven.compiler.source=1.8.

Since the warnings are treated as errors due to -Werror, the build
fails. We should set <showWarnings>true</showWarnings> once we bump the
maven.compiler.source.

Another change was to remove the deprecated
<failBuildOnAnyVulnerability>true</failBuildOnAnyVulnerability>
in favor of
<failBuildOnCVSS>0.0</failBuildOnCVSS>