Use writeTo() to buffer HttpEntity in httpclient
With a change in Spring Framework 6.1, HttpComponentsClientHttpRequest has an internal class BodyEntity, that is meant to not cache an `OutputStream`, but to provide a callback, that's used when the `OutputStream` is required. The `getContent()` method throws `UnsupportedOperationException`.

With this change, I'm using the `writeTo()` method to copy the `OutputStream`, as for Logbook, buffering the stream is required.