Attempt to write a response log when flushBuffer() fails (#1997)
* Attempt to write a response log when flushBuffer() fails

In cases when a reactive client closes a connection, flushBuffer() operation may throw a "java.io.IOException: Broken pipe" exception, that in its own turn leads to a connection close, and `LogbookAsyncListener.onComplete` call. But as this happens AFTER `ServletResponse.freeResources()` is called (by HttpServerExchange.endExchange() in case of undertow, for example), the `responseWritingStageSynchronizationName` attribute is not present anymore in the response object, which leads to an NPE.
This fix attempts to overcome this edge case, by trying to write the response anyway, even after the flush fails.

* add isolated unit test