Check for synchronization attribute to not be null
In rare cases when an async flow is enabled, LogbookFilter throws NPE because the synchronization attribute is missing in the request context. The root cause for such scenarios is not clear and happens rarely.
The hypothesis is that it happens due to possible race conditions that happen when the calling client cancels the request and closes the connection. But it wasn't reproduced locally.

With this change the logging of the request & response will not happen.