Use default coroutine context in ktor client
It appears that when nestin coroutine contexts via `scope.launch(coroutineContext) {}` LogbookClient may lead to deadlocks on big responses, as `loggingContent.discard()` uses blocking operations underneath and if other plugins interferred with the current coroutine scope.

When inheriting the scope it seems that the issue is resolved, as Dispatchers.IO is used and not kotlinx.coroutines.BlockingEventLoop as the coroutine element.