Support entityDetails with negative content length (#2143)
* Support entityDetails with negative content length

In case of chunked transfer encoding, org.apache.hc.core5.http.impl.DefaultContentLengthStrategy uses -1 as the content length of a org.apache.hc.core5.http.HttpMessage. Which later passed in org.apache.hc.core5.http.EntityDetails. This is later causes an `java.lang.NegativeArraySizeException: -1` exception as Logobook tried to determine the buffer size for the body.

The suggested fix falls back to java.nio.Buffer.remaining() to determine the size of the buffer, when Content Length is negative.

* apply the fallback only for -1 length (CHUNK strategy)